+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Adds additional configurations for `CalendarPageHeader`, `MonthPageHeader`, `DayPageHeader` and `WeekPageHeader`.
     - Added `titleBuilder` to build custom title for header.
 - Fixes issue calendar scroll physics for day & week view. [#417](https://github.com/SimformSolutionsPvtLtd/flutter_calendar_view/issues/417)
+- Adds `onTimestampTap` callback in `WeekView`
+  and `DayView`.  [#383](https://github.com/SimformSolutionsPvtLtd/flutter_calendar_view/issues/383)
 - `Deprecations`:
     - deprecated `backgroundColor` and `iconColor` from `CalendarPageHeader`, `DayPageHeader`, `MonthPageHeader` and `WeekPageHeader`.
         - **Solution:** use `headerStyle` instead.